+++ patterns/fizzy_lifting_drink.py
@@ -12,8 +12,8 @@ class FizzyLiftingDrink(Pattern):
         self.cache_positions(leds)
 
     def cache_positions(self, leds):
-        self._led_deltas = np.array(list(led.coord.get_delta("global") for led in leds), dtype=np.float16)
-        self._led_thetas = np.array(list(led.coord.get("global", "spherical").theta for led in leds), dtype=np.float16)
+        self._led_deltas = np.array(list(led.coord.get_delta("global") for led in leds), dtype=np.float32)
+        self._led_thetas = np.array(list(led.coord.get("global", "spherical").theta for led in leds), dtype=np.float32)
 
     def get_pixel_colours(self, leds, time, palette_handler, palette_name):
         # how many sine wave cycles are squeezed into our n_pixels
@@ -39,13 +39,13 @@ class FizzyLiftingDrink(Pattern):
         g = blackstripes * np.cos(pct * freq_g + t / speed_g)
         b = blackstripes * np.cos(pct * freq_b + t / speed_b)
 
-        # output = np.zeros((len(leds), 3), dtype=np.float16)
+        # output = np.zeros((len(leds), 3), dtype=np.float32)
         # np.put_along_axis(output, [[0]] * num_leds, r, 1)
         # np.put_along_axis(output, [1], g, 1)
         # np.put_along_axis(output, [2], b, 1)
         # output = np.array(list([r[i], g[i], b[i]] for i in range(len(leds))))
 
-        output = np.zeros((len(leds), 3), dtype=np.float16)
+        output = np.zeros((len(leds), 3), dtype=np.float32)
         output[..., 0] = r
         output[..., 1] = g
         output[..., 2] = b
@@ -53,7 +53,7 @@ class FizzyLiftingDrink(Pattern):
         output += 1
         output *= 127
 
-        raw_colours = palette_handler.sample_radial_all(self._led_deltas, time, self._space_factor, self._time_factor, palette_name).astype(np.float16)
+        raw_colours = palette_handler.sample_radial_all(self._led_deltas, time, self._space_factor, self._time_factor, palette_name).astype(np.float32)
         raw_colours *= blackstripes[:,np.newaxis]
 
         return output * 0.5 + raw_colours * 0.5
